+++ ttnn/cpp/ttnn/operations/conv/conv2d/prepare_conv2d_weights.cpp
@@ -1152,66 +1152,25 @@ static Conv2dWeightsBiasPrepConfig setup_conv_prep_config(
     is_dram_conv = is_dram_conv && !is_conv1d;
 
     if (is_dram_conv) {
-        Conv2dSliceConfig dram_slice_config;
-        if (dram_slice_config_.has_value()) {
-            dram_slice_config = dram_slice_config_.value();
-        } else {
-            Tensor dummy_weight_tensor = tt::tt_metal::create_device_tensor(
-                tt::tt_metal::TensorSpec(
-                    ttnn::Shape({out_channels, in_channels / groups, kernel_size[0], kernel_size[1]}),
-                    tt::tt_metal::TensorLayout(
-                        conv_config.weights_dtype.value(),
-                        tt::tt_metal::PageConfig(Layout::ROW_MAJOR),
-                        MemoryConfig{
-                            TensorMemoryLayout::INTERLEAVED,
-                            BufferType::DRAM,
-                        })),
-                device);
-            std::optional<Tensor> dummy_bias_tensor = std::nullopt;
-            if (has_bias) {
-                dummy_bias_tensor = tt::tt_metal::create_device_tensor(
-                    tt::tt_metal::TensorSpec(
-                        ttnn::Shape({1, 1, 1, out_channels}),
-                        tt::tt_metal::TensorLayout(
-                            conv_config.weights_dtype.value(),
-                            tt::tt_metal::PageConfig(Layout::ROW_MAJOR),
-                            MemoryConfig{
-                                TensorMemoryLayout::INTERLEAVED,
-                                BufferType::DRAM,
-                            })),
-                    device);
-            }
-            auto op_slice_attr = Conv2dSliceAttr(
-                batch_size,
-                {input_height, input_width},
-                in_channels,
-                out_channels,
-                kernel_size,
-                stride,
-                padding_n4,
-                dilation,
-                groups,
-                input_layout,
-                input_dtype,
-                conv_output_dtype,
-                std::ref(dummy_weight_tensor),
-                has_bias ? std::make_optional(std::ref(dummy_bias_tensor.value())) : std::nullopt,
-                conv_config,
-                compute_config,
-                device);
-            dram_slice_config = determine_slice_config(
-                &op_slice_attr,
-                ttnn::Shape{batch_size, input_height, input_width, in_channels},
-                ttnn::Shape{batch_size, output_height, output_width, out_channels},
-                dram_slice_config_,
-                conv_config.output_layout,
-                device);
-            log_info(
-                tt::LogOp,
-                "Auto determined DRAM Slice Config in Prepare Conv2d Weights as {} for {}",
-                dram_slice_config,
-                op_slice_attr.name());
-        }
+        Conv2dSliceConfig dram_slice_config = determine_dram_slice_config_for_prepare_weights(
+            batch_size,
+            input_height,
+            input_width,
+            in_channels,
+            out_channels,
+            kernel_size,
+            stride,
+            padding_n4,
+            dilation,
+            groups,
+            input_layout,
+            input_dtype,
+            conv_output_dtype,
+            has_bias,
+            conv_config,
+            compute_config,
+            device,
+            dram_slice_config_);
         uint32_t slice_rounding_value = 1;
         if (conv_config.output_layout == tt_metal::Layout::TILE &&
             dram_slice_config.slice_type == Conv2dSliceConfig::SliceType::DRAM_WIDTH) {
